Rename shared styles file
