+++ static/shared.ts
